# backend/models/chat_message_request.go
// Code generated by go-swagger; DO NOT EDIT.

package models

import (
	"context"
	"encoding/json"

	"github.com/go-openapi/errors"
	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
	"github.com/go-openapi/validate"
)

// ChatMessageRequest chat message request
//
// swagger:model ChatMessageRequest
type ChatMessageRequest struct {

	// Base64 payload of image/audio/other file (optional)
	AttachmentBase64 string `json:"attachment_base64,omitempty"`

	// MIME type
	AttachmentMimeType string `json:"attachment_mime_type,omitempty"`

	// Original filename for attachment
	AttachmentName string `json:"attachment_name,omitempty"`

	// attachment type
	// Enum: ["image","audio","file"]
	AttachmentType string `json:"attachment_type,omitempty"`

	// content
	// Min Length: 1
	Content string `json:"content,omitempty"`
}

// Validate validates this chat message request
func (m *ChatMessageRequest) Validate(formats strfmt.Registry) error {
	var res []error

	if err := m.validateAttachmentType(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateContent(formats); err != nil {
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

var chatMessageRequestTypeAttachmentTypePropEnum []any

func init() {
	var res []string
	if err := json.Unmarshal([]byte(`["image","audio","file"]`), &res); err != nil {
		panic(err)
	}
	for _, v := range res {
		chatMessageRequestTypeAttachmentTypePropEnum = append(chatMessageRequestTypeAttachmentTypePropEnum, v)
	}
}

const (

	// ChatMessageRequestAttachmentTypeImage captures enum value "image"
	ChatMessageRequestAttachmentTypeImage string = "image"

	// ChatMessageRequestAttachmentTypeAudio captures enum value "audio"
	ChatMessageRequestAttachmentTypeAudio string = "audio"

	// ChatMessageRequestAttachmentTypeFile captures enum value "file"
	ChatMessageRequestAttachmentTypeFile string = "file"
)

// prop value enum
func (m *ChatMessageRequest) validateAttachmentTypeEnum(path, location string, value string) error {
	if err := validate.EnumCase(path, location, value, chatMessageRequestTypeAttachmentTypePropEnum, true); err != nil {
		return err
	}
	return nil
}

func (m *ChatMessageRequest) validateAttachmentType(formats strfmt.Registry) error {
	if swag.IsZero(m.AttachmentType) { // not required
		return nil
	}

	// value enum
	if err := m.validateAttachmentTypeEnum("attachment_type", "body", m.AttachmentType); err != nil {
		return err
	}

	return nil
}

func (m *ChatMessageRequest) validateContent(formats strfmt.Registry) error {
	if swag.IsZero(m.Content) { // not required
		return nil
	}

	if err := validate.MinLength("content", "body", m.Content, 1); err != nil {
		return err
	}

	return nil
}

// ContextValidate validates this chat message request based on context it is used
func (m *ChatMessageRequest) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *ChatMessageRequest) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *ChatMessageRequest) UnmarshalBinary(b []byte) error {
	var res ChatMessageRequest
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}