# openapi.yaml
openapi: 3.0.0
info:
  title: Medbratishka API
  description: API платформы для связи между докторами и пациентами
  version: 1.0.0
  contact:
    name: API Support
    url: https://medbratishka.com/support
    email: support@medbratishka.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.medbratishka.com
    description: Production server
paths:
...
  /chats:
    get:
      summary: Get my chats
      tags:
        - Chat
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Chats list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatsResponse'
    post:
      summary: Create chat with doctor
      tags:
        - Chat
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChatRequest'
      responses:
        '200':
          description: Chat created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /chats/{chat_id}/messages:
    get:
      summary: Get chat messages
      tags:
        - Chat
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: chat_id
          required: true
          schema:
            type: integer
            format: int64
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Messages list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatMessagesResponse'
    post:
      summary: Send message to chat
      tags:
        - Chat
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: chat_id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatMessageRequest'
      responses:
        '200':
          description: Message sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatMessageResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: Код ошибки
          example: UNAUTHORIZED
        message:
          type: string
          description: Сообщение об ошибке
          example: invalid credentials
        type:
          type: string
          description: Тип ошибки
          example: unauthorized
        details:
          type: object
          description: Детали ошибки
          nullable: true
...
    CreateChatRequest:
      type: object
      required:
        - doctor_id
      properties:
        doctor_id:
          type: integer
          format: int64
    ChatMessageRequest:
      type: object
      properties:
        content:
          type: string
          minLength: 1
        attachment_base64:
          type: string
          description: Base64 payload of image/audio/other file (optional)
        attachment_name:
          type: string
          description: Original filename for attachment
        attachment_mime_type:
          type: string
          description: MIME type
        attachment_type:
          type: string
          enum:
            - image
            - audio
            - file
    ChatMessageResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        sender_id:
          type: integer
          format: int64
        sender_name:
          type: string
        content:
          type: string
        attachment_url:
          type: string
        attachment_name:
          type: string
        attachment_type:
          type: string
        attachment_mime_type:
          type: string
        created_at:
          type: integer
          format: int64
    ChatResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        doctor_id:
          type: integer
          format: int64
        patient_id:
          type: integer
          format: int64
        other_user_id:
          type: integer
          format: int64
        other_login:
          type: string
        other_name:
          type: string
        last_message_id:
          type: integer
          format: int64
        last_message:
          type: string
        last_message_at:
          type: integer
          format: int64
        unread_count:
          type: integer
          format: int64
        last_read_message_id:
          type: integer
          format: int64
        has_unread:
          type: boolean
        updated_at:
          type: integer
          format: int64
    ChatsResponse:
      type: object
      properties:
        chats:
          type: array
          items:
            $ref: '#/components/schemas/ChatResponse'
    ChatMessagesResponse:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: '#/components/schemas/ChatMessageResponse'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен для аутентификации
...